+++ lib/globals.dart
@@ -20,8 +20,8 @@ final GlobalKey<ScaffoldMessengerState> scaffoldMessengerKey =
 final GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
 
 const String prgName = 'easysend';
-const String progVersion = '0.7.20260904';
-const int buildNumber = 146;
+const String progVersion = '0.8.20260904';
+const int buildNumber = 148;
 const String progAuthor = 'Eugen';
 
 const String langFile = 'assets/locales.json';